requirements.txt
alembic==1.0.8
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.6.1
